Some General Tips About Solana Development
In the Solana Development, you will face a lot of weird custom errors bugs, and it can be quite hard and frustrating to fix them since the Solana dev ecosystem is not as big as the Egg dev ecosystem. But not to worry, when you get stuck, you simply have to look in the right place for the solution.